// client/src/Gen/models/MealRecipeEntryPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nutrition Tracker API
 * API Specification for Nutrition Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface MealRecipeEntryPostRequest
 */
export interface MealRecipeEntryPostRequest {
    /**
     *
     * @type {number}
     * @memberof MealRecipeEntryPostRequest
     */
    amount: number;
    /**
     *
     * @type {number}
     * @memberof MealRecipeEntryPostRequest
     */
    recipeId: number;
}

/**
 * Check if a given object implements the MealRecipeEntryPostRequest interface.
 */
export function instanceOfMealRecipeEntryPostRequest(value: object): value is MealRecipeEntryPostRequest {
    if (!("amount" in value) || value["amount"] === undefined) return false;
    if (!("recipeId" in value) || value["recipeId"] === undefined) return false;
    return true;
}

export function MealRecipeEntryPostRequestFromJSON(json: any): MealRecipeEntryPostRequest {
    return MealRecipeEntryPostRequestFromJSONTyped(json, false);
}

export function MealRecipeEntryPostRequestFromJSONTyped(
    json: any,
    ignoreDiscriminator: boolean,
): MealRecipeEntryPostRequest {
    if (json == null) {
        return json;
    }
    return {
        amount: json["amount"],
        recipeId: json["recipeId"],
    };
}

export function MealRecipeEntryPostRequestToJSON(json: any): MealRecipeEntryPostRequest {
    return MealRecipeEntryPostRequestToJSONTyped(json, false);
}

export function MealRecipeEntryPostRequestToJSONTyped(
    value?: MealRecipeEntryPostRequest | null,
    ignoreDiscriminator: boolean = false,
): any {
    if (value == null) {
        return value;
    }

    return {
        amount: value["amount"],
        recipeId: value["recipeId"],
    };
}