// client/src/Gen/models/MealRecipeEntryResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Nutrition Tracker API
 * API Specification for Nutrition Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface MealRecipeEntryResponse
 */
export interface MealRecipeEntryResponse {
    /**
     *
     * @type {number}
     * @memberof MealRecipeEntryResponse
     */
    id: number;
    /**
     *
     * @type {number}
     * @memberof MealRecipeEntryResponse
     */
    amount: number;
    /**
     *
     * @type {number}
     * @memberof MealRecipeEntryResponse
     */
    sequenceNumber: number;
    /**
     *
     * @type {number}
     * @memberof MealRecipeEntryResponse
     */
    recipeId: number;
}

/**
 * Check if a given object implements the MealRecipeEntryResponse interface.
 */
export function instanceOfMealRecipeEntryResponse(value: object): value is MealRecipeEntryResponse {
    if (!("id" in value) || value["id"] === undefined) return false;
    if (!("amount" in value) || value["amount"] === undefined) return false;
    if (!("sequenceNumber" in value) || value["sequenceNumber"] === undefined) return false;
    if (!("recipeId" in value) || value["recipeId"] === undefined) return false;
    return true;
}

export function MealRecipeEntryResponseFromJSON(json: any): MealRecipeEntryResponse {
    return MealRecipeEntryResponseFromJSONTyped(json, false);
}

export function MealRecipeEntryResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): MealRecipeEntryResponse {
    if (json == null) {
        return json;
    }
    return {
        id: json["id"],
        amount: json["amount"],
        sequenceNumber: json["sequenceNumber"],
        recipeId: json["recipeId"],
    };
}

export function MealRecipeEntryResponseToJSON(json: any): MealRecipeEntryResponse {
    return MealRecipeEntryResponseToJSONTyped(json, false);
}

export function MealRecipeEntryResponseToJSONTyped(
    value?: MealRecipeEntryResponse | null,
    ignoreDiscriminator: boolean = false,
): any {
    if (value == null) {
        return value;
    }

    return {
        id: value["id"],
        amount: value["amount"],
        sequenceNumber: value["sequenceNumber"],
        recipeId: value["recipeId"],
    };
}